+++ packages/api-client/package.json
@@ -83,6 +83,11 @@
       "types": "./dist/views/Request/RequestSection/RequestAuth/index.d.ts",
       "default": "./dist/views/Request/RequestSection/RequestAuth/index.js"
     },
+    "./views/Components/CodeSnippet": {
+      "import": "./dist/views/Components/CodeSnippet/index.js",
+      "types": "./dist/views/Components/CodeSnippet/index.d.ts",
+      "default": "./dist/views/Components/CodeSnippet/index.js"
+    },
     "./types": {
       "import": "./dist/types/index.js",
       "types": "./dist/types/index.d.ts",
