+++ packages/api-client/src/views/Components/CodeSnippet/index.ts
@@ -1,0 +1,1 @@
+export { default as CodeSnippet } from './CodeSnippet.vue'
